+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.5] - 2026-18-06
+
+### Changed
+
+* Updated CHANGELOG.md.
+* Extending valid search chains of classes VsphereDatastoreDict and VsphereDsClusterDict.
+* Added classmethod `valid_search_chains()` to classes VsphereDatastoreDict
+  and VsphereDsClusterDict.
+
+## [1.8.4] - 2026-02-06
+
+### Changed
+
+* Non change release for triggering build pipelines.
+
 ## [1.8.3] - 2026-01-22
 
 ### Fixed
@@ -295,6 +310,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Initial release.
 
 
+[1.8.5]: https://github.com/fbrehm/fb-vmware/compare/1.8.4...1.8.5
+[1.8.4]: https://github.com/fbrehm/fb-vmware/compare/1.8.3...1.8.4
+[1.8.3]: https://github.com/fbrehm/fb-vmware/compare/1.8.2...1.8.3
+[1.8.2]: https://github.com/fbrehm/fb-vmware/compare/1.8.1...1.8.2
+[1.8.1]: https://github.com/fbrehm/fb-vmware/compare/1.8.0...1.8.1
+[1.8.0]: https://github.com/fbrehm/fb-vmware/compare/1.7.1...1.8.0
+[1.7.1]: https://github.com/fbrehm/fb-vmware/compare/1.7.0...1.7.1
 [1.7.0]: https://github.com/fbrehm/fb-vmware/compare/1.6.0...1.7.0
 [1.6.0]: https://github.com/fbrehm/fb-vmware/compare/1.5.3...1.6.0
 [1.5.3]: https://github.com/fbrehm/fb-vmware/compare/1.5.2...1.5.3